сделала правки и первую часть домашки
